# api/openapi/contract.base.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "CPRa SDK draft management contract",
    "version": "0.1.0-rc.1",
    "description": "Draft contract. V2 server qualification is pending. Existing v1 is separate."
  },
  "paths": {
...
    "/api/v2/operations": {
      "post": {
        "operationId": "CreateOperation",
        "summary": "CreateOperation",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./models.json#/components/schemas/OperationCreateRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "ListOperations",
        "summary": "ListOperations",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/OperationList"
                }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "selector",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "monitorID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v2/operations/{id}": {
      "get": {
        "operationId": "GetOperation",
        "summary": "GetOperation",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4096
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "description": "Read the original operation and, when executionResult.state is ready, one immutable original-input-order result page. Cursor and limit bind the retained execution result; they do not activate or cancel the operation."
      }
    },
...
  }
}